+++ src/main/java/dev/orne/config/commons/prefs/PreferencesBuilderParameters.java
@@ -1,0 +1,47 @@
+package dev.orne.config.commons.prefs;
+
+/*-
+ * #%L
+ * Orne Config
+ * %%
+ * Copyright (C) 2019 - 2021 Orne Developments
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-3.0.html>.
+ * #L%
+ */
+
+import org.apache.commons.configuration2.builder.BuilderParameters;
+import org.apache.commons.configuration2.builder.HierarchicalBuilderProperties;
+import org.apache.commons.configuration2.tree.ImmutableNode;
+
+/**
+ * Definition of a parameters interface providing a fluent API for setting all
+ * properties for a preferences based configuration.
+ * <p>
+ * <strong>Important note:</strong> This interface is not intended to be
+ * implemented by client code! It defines a set of available properties and may
+ * be extended even in minor releases.
+ *
+ * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+ * @version 1.0, 2021-01
+ * @param <N> The type of nodes of the generated {@code PreferencesBased}
+ * @since 0.2
+ */
+public interface PreferencesBuilderParameters
+extends PreferencesBuilderProperties<PreferencesBuilderParameters, ImmutableNode>,
+        HierarchicalBuilderProperties<PreferencesBuilderParameters>,
+        BuilderParameters {
+    // No extra methods
+}